crypto/requirements.txt
fastapi
pyjwt==2.4.0
pydantic
sqlalchemy
uvicorn>=0.15.0,<0.16.0
